Software that allows camping officials to easily manage activities
This is a software dedicated to icamp, it allows users to manage
activities conveiniently on their computers instead of having to write
down with pen and paper. This project makes use of a MySQL database
which is stored offline so ensure you have your Wamp server setup and
update the connection string if need be and finally import our SQL file.